requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
paho-mqtt==2.0.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.1
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
